move signature verification tests to core
