+++ eth2util/signing/signing_test.go
@@ -23,95 +23,13 @@ import (
 	"testing"
 
 	eth2v1 "github.com/attestantio/go-eth2-client/api/v1"
-	eth2p0 "github.com/attestantio/go-eth2-client/spec/phase0"
-	"github.com/coinbase/kryptology/pkg/signatures/bls/bls_sig"
 	"github.com/stretchr/testify/require"
 
-	"github.com/obolnetwork/charon/core"
 	"github.com/obolnetwork/charon/eth2util/signing"
 	"github.com/obolnetwork/charon/tbls"
 	"github.com/obolnetwork/charon/tbls/tblsconv"
-	"github.com/obolnetwork/charon/testutil"
 	"github.com/obolnetwork/charon/testutil/beaconmock"
 )
-
-func TestVerifySignedData(t *testing.T) {
-	tests := []struct {
-		name   string
-		data   core.Eth2SignedData
-		domain signing.DomainName
-	}{
-		{
-			name:   "verify attestation",
-			data:   core.NewAttestation(testutil.RandomAttestation()),
-			domain: signing.DomainBeaconAttester,
-		},
-		{
-			name:   "verify beacon block",
-			data:   testutil.RandomCoreVersionSignedBeaconBlock(t),
-			domain: signing.DomainBeaconProposer,
-		},
-		{
-			name:   "verify blinded beacon block",
-			data:   testutil.RandomCoreVersionSignedBlindedBeaconBlock(t),
-			domain: signing.DomainBeaconProposer,
-		},
-		{
-			name:   "verify randao",
-			data:   testutil.RandomCoreSignedRandao(),
-			domain: signing.DomainRandao,
-		},
-		{
-			name:   "verify voluntary exit",
-			data:   core.NewSignedVoluntaryExit(testutil.RandomExit()),
-			domain: signing.DomainExit,
-		},
-		{
-			name:   "verify registration",
-			data:   testutil.RandomCoreVersionedSignedValidatorRegistration(t),
-			domain: signing.DomainApplicationBuilder,
-		},
-		{
-			name:   "verify beacon committee selection",
-			data:   testutil.RandomCoreBeaconCommitteeSelection(),
-			domain: signing.DomainSelectionProof,
-		},
-		{
-			name: "verify attestation aggregate and proof",
-			data: core.SignedAggregateAndProof{
-				SignedAggregateAndProof: eth2p0.SignedAggregateAndProof{
-					Message: testutil.RandomAggregateAndProof(),
-				},
-			},
-			domain: signing.DomainAggregateAndProof,
-		},
-		{
-			name:   "verify sync committee message",
-			data:   core.NewSignedSyncMessage(testutil.RandomSyncCommitteeMessage()),
-			domain: signing.DomainSyncCommittee,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			bmock, err := beaconmock.New()
-			require.NoError(t, err)
-
-			epoch, err := test.data.Epoch(context.Background(), bmock)
-			require.NoError(t, err)
-
-			root, err := test.data.MessageRoot()
-			require.NoError(t, err)
-
-			sigData, err := signing.GetDataRoot(context.Background(), bmock, test.domain, epoch, root)
-			require.NoError(t, err)
-
-			sig, pubkey := sign(t, sigData[:])
-
-			require.NoError(t, signing.Verify(context.Background(), bmock, test.domain, epoch, root, sig, pubkey))
-		})
-	}
-}
 
 func TestVerifyRegistrationReference(t *testing.T) {
 	bmock, err := beaconmock.New()
@@ -154,16 +72,10 @@ func TestVerifyRegistrationReference(t *testing.T) {
 		fmt.Sprintf("%x", registration.Signature),
 		fmt.Sprintf("%x", sigEth2),
 	)
-}
 
-func sign(t *testing.T, data []byte) (eth2p0.BLSSignature, *bls_sig.PublicKey) {
-	t.Helper()
-
-	pk, secret, err := tbls.Keygen()
+	pubkey, err := secretShare.GetPublicKey()
 	require.NoError(t, err)
 
-	sig, err := tbls.Sign(secret, data)
+	err = signing.Verify(context.Background(), bmock, signing.DomainApplicationBuilder, 0, sigRoot, tblsconv.SigToETH2(sig), pubkey)
 	require.NoError(t, err)
-
-	return tblsconv.SigToETH2(sig), pk
 }
